Complete merge for Custom Post Types filter.
Includes whitespace fixes and new filter type for custom post types.

This version of the code doesn't make a big deal about the fact that
they are custom post types. It simply lists the post types (including
the built-in 'Blog Post' and any custom post types that have been
defined).